refactor : 자동차 이름도 원시값 포장하여 책임 분리
