+++ src/main/java/racingcar/MainController.java
@@ -4,6 +4,7 @@ import static racingcar.domain.GameOption.CAR_NAME_DELIMITER;
 
 import java.util.List;
 import racingcar.domain.Car;
+import racingcar.domain.CarName;
 import racingcar.domain.Cars;
 import racingcar.domain.Position;
 import racingcar.domain.TrialCount;
@@ -24,7 +25,8 @@ public class MainController {
         String inputValue = InputView.inputString(ViewMessage.INPUT_CAR_NAME);
         List<Car> carList = Converter.splitByDelimiter(inputValue, CAR_NAME_DELIMITER)
                 .stream()
-                .map(name -> new Car(name, new Position(0)))
+                .map(CarName::new)
+                .map(carName -> new Car(carName, new Position(0)))
                 .toList();
         return new Cars(carList);
     }
