+++ src/test/java/racingcar/CarTest.java
@@ -4,6 +4,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import racingcar.domain.Car;
+import racingcar.domain.CarName;
 import racingcar.domain.Position;
 import racingcar.message.ErrorMessage;
 
@@ -12,7 +13,7 @@ public class CarTest {
     @ParameterizedTest
     @ValueSource(strings = {"a", "pobi", "a2", "abcd", "1234"})
     void CarNameSuccessTest(String name) {
-        Car car = new Car(name, new Position(0));
+        Car car = new Car(new CarName(name), new Position(0));
 
         Assertions.assertThat(car.getName()).isEqualTo(name);
     }
@@ -21,7 +22,7 @@ public class CarTest {
     @ValueSource(strings = {"", "         ", "12345", "thisIsMoreThenFour"})
     void CarNameFailTest(String name) {
         Assertions.assertThatThrownBy(() -> {
-                    new Car(name, new Position(0));
+                    new Car(new CarName(name), new Position(0));
                 }).isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(ErrorMessage.INVALID_CAR_NAME);
     }
