+++ src/test/java/racingcar/RacingServiceTest.java
@@ -6,6 +6,7 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racingcar.domain.Car;
+import racingcar.domain.CarName;
 import racingcar.domain.Cars;
 import racingcar.domain.Position;
 
@@ -14,7 +15,7 @@ public class RacingServiceTest {
     @DisplayName("자동차가 1회 움직이면 위치는 1 이다.")
     @Test
     void moveSuccessTest() {
-        Car car = new Car("car", new Position(0));
+        Car car = new Car(new CarName("car"), new Position(0));
         car.moveForward();
         assertThat(car.getPosition()).isEqualTo(1);
     }
@@ -22,16 +23,16 @@ public class RacingServiceTest {
     @DisplayName("움직이지 않으면 위치는 0 이다.")
     @Test
     void moveFailTest() {
-        Car car = new Car("car", new Position(0));
+        Car car = new Car(new CarName("car"), new Position(0));
         assertThat(car.getPosition()).isEqualTo(0);
     }
 
     @DisplayName("우승자가 1명만 있는 경우 리스트에는 한 명만 있다.")
     @Test
     void getSingleWinnerTest() {
-        Car car1 = new Car("car1", new Position(0));
-        Car car2 = new Car("car2", new Position(0));
-        Car car3 = new Car("car3", new Position(0));
+        Car car1 = new Car(new CarName("car1"), new Position(0));
+        Car car2 = new Car(new CarName("car2"), new Position(0));
+        Car car3 = new Car(new CarName("car3"), new Position(0));
 
         car1.moveForward();
         List<Car> carList = List.of(car1, car2, car3);
@@ -44,9 +45,9 @@ public class RacingServiceTest {
     @DisplayName("우승자가 2명이 있는 경우 리스트에는 2명만 있다.")
     @Test
     void getDoubleWinnerTest() {
-        Car car1 = new Car("car1", new Position(0));
-        Car car2 = new Car("car2", new Position(0));
-        Car car3 = new Car("car3", new Position(0));
+        Car car1 = new Car(new CarName("car1"), new Position(0));
+        Car car2 = new Car(new CarName("car2"), new Position(0));
+        Car car3 = new Car(new CarName("car3"), new Position(0));
 
         car1.moveForward();
         car2.moveForward();
